refactor: Export code from ContactsListModal into ContactsListContent
no code change
